+++ src/Rose/BinaryAnalysis/InstructionSemantics/SymbolicSemantics.C
@@ -2108,14 +2108,14 @@ RiscOperators::writeMemory(RegisterDescriptor segreg,
 }
 
 BaseSemantics::SValue::Ptr
-RiscOperators::readLocal(uint8_t index) {
+RiscOperators::readLocal(size_t index) {
     auto frame = currentState()->frameState();
     ASSERT_not_null(frame);
     return frame->readLocal(index);
 }
 
 void
-RiscOperators::writeLocal(uint8_t index, const BaseSemantics::SValue::Ptr &value){
+RiscOperators::writeLocal(size_t index, const BaseSemantics::SValue::Ptr &value){
     auto frame = currentState()->frameState();
     ASSERT_not_null(frame);
     ASSERT_not_null(value);
